smart_aleck/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
Django==5.1.1
django-sslserver==0.22
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
mysqlclient==2.2.4
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
openai>=1.0.0
pinecone-client>=3.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
torch>=2.0.0
